+++ server/src/api/user-inventory/user-inventory.service.ts
@@ -97,7 +97,8 @@ export class UserInventoryService {
 			});
 
 			return {
-				items: mappedItems,
+				// items: mappedItems,
+				items: [],
 				total,
 			};
 		});